fix: review fixes — CLI edge cases, duplicate React keys, DRY plugin updates hook
- Fix toPackageName stripping @ and / from allowed chars, fallback to
  default name for degenerate inputs
- Read CLI version from package.json instead of hardcoding
- Fix duplicate React key warning in profanity filter word list
- Extract ONE_HOUR and usePluginUpdates to shared helpers module
