+++ .lore.md
@@ -363,6 +363,9 @@
 <!-- lore:019ee643-62b7-711f-855a-94cdc4a116ee -->
 * **Always request read-only architectural audits with exact file paths and line numbers before making changes**: Before implementing any feature or fix, the user consistently reads the full contents of relevant files with line numbers. They typically request multiple related files at once (2-4 files per request), specifying exact file paths relative to the project root. They want complete file contents or specific line ranges, not summaries. This applies to any file they plan to modify or that contains logic relevant to the task. The pattern suggests they want to fully understand the existing code structure, exports, and dependencies before proposing or writing any changes.
 
+<!-- lore:019f21e6-e991-7369-b3e4-09b8f48c900c -->
+* **Always request thorough code review with specific focus areas before merging branches**: When working on feature branches, the user requests comprehensive code reviews comparing the branch to main before merging. Reviews consistently cover: (1) bugs — logic errors, race conditions, null checks, type issues; (2) DRY violations — duplicate code across files; (3) inconsistencies — naming mismatches, missing error handling, hardcoded values; (4) missing edge cases — empty inputs, special characters, boundary conditions. The user also provides context about which duplications are intentional (e.g., standalone templates). When conducting such reviews, cover all changed files systematically and flag issues in each category explicitly.
+
 <!-- lore:019f1703-8c87-7813-90aa-d41f8432d67b -->
 * **Always request thorough file audits with structured output before starting migrations**: Before beginning a migration or refactor effort, the user consistently requests a comprehensive scan/audit of all relevant files in a directory. They define specific patterns to search for, want results organized per-file with metadata (file path, line numbers, pattern description, classification, complexity rating), and want results ranked or prioritized. This applies to React Query migrations, plugin system changes, and similar cross-cutting refactors. When responding, proactively read all relevant files in parallel, classify findings against the user's defined criteria, and present structured output with rankings before proposing any implementation plan.
