+++ apps/electron/src/renderer/src/pages/plugins/helpers.ts
@@ -1,5 +1,9 @@
-import type { PluginInfo } from "@shared/plugins";
+import type { PluginInfo, PluginUpdateResult } from "@shared/plugins";
+import { useQuery } from "@tanstack/react-query";
 import { type LucideIcon, icons as lucideIcons, Puzzle } from "lucide-react";
+import { useMemo } from "react";
+
+export const ONE_HOUR = 60 * 60 * 1000;
 
 /**
  * Resolve a lucide icon by name, accepting PascalCase (`FileMusic`) or
@@ -29,4 +33,30 @@ export function pluginDisplayName(plugin: PluginInfo): string {
     .filter(Boolean)
     .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
     .join(" ");
+}
+
+/**
+ * Shared hook for checking plugin updates. Uses a stable query key family so
+ * caching is shared between the plugins list and detail pages.
+ */
+export function usePluginUpdates(plugins: PluginInfo[]) {
+  const entries = useMemo(
+    () =>
+      plugins
+        .filter((p) => p.version && !p.missing)
+        .map((p) => ({ name: p.specifier, currentVersion: p.version! })),
+    [plugins],
+  );
+
+  return useQuery({
+    queryKey: ["plugin-updates", entries],
+    queryFn: async () => {
+      if (entries.length === 0) return new Map<string, PluginUpdateResult>();
+      const results = await window.api.checkPluginUpdates(entries);
+      return new Map(results.map((r) => [r.name, r]));
+    },
+    staleTime: ONE_HOUR,
+    retry: 1,
+    enabled: entries.length > 0,
+  });
 }
